hashing: make new hashing work on 32-bit systems.

+++ base/pkg/types.jl
@@ -44,7 +44,7 @@ function intersect(A::VersionSet, B::VersionSet)
     VersionSet(ivals)
 end
 isequal(A::VersionSet, B::VersionSet) = (A.intervals == B.intervals)
-hash(s::VersionSet, h::Uint) = hash(s.intervals, h + 0x2fd2ca6efa023f44)
+hash(s::VersionSet, h::Uint) = hash(s.intervals, h + uint(0x2fd2ca6efa023f44))
 deepcopy_internal(vs::VersionSet, ::ObjectIdDict) = VersionSet(copy(vs.intervals))
 
 typealias Requires Dict{ByteString,VersionSet}
